moved jwttoken parser
