+++ source/Api/Extensions/DependencyInjection/B2BAuthenticationExtensions.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.EDI.Api.Authentication;
 using Energinet.DataHub.EDI.Api.Authentication.Certificate;
-using Energinet.DataHub.EDI.Api.Configuration.Authentication;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Authentication;
 using Energinet.DataHub.EDI.MasterData.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
