+++ source/IntegrationTests/Api/WhenMarketActorAuthenticatorMiddlewareIsCalledTests.cs
@@ -20,7 +20,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.Core.App.FunctionApp;
-using Energinet.DataHub.EDI.Api.Configuration.Authentication;
+using Energinet.DataHub.EDI.Api.Authentication;
 using Energinet.DataHub.EDI.Api.Configuration.Middleware.Authentication;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
